+++ R/tableby.internal.R
@@ -38,6 +38,10 @@ format.tbstat <- function(x, digits = NULL, ...)
 {
   x <- x[] # to remove classes
   if(is.numeric(x)) x <- trimws(formatC(x, digits = digits, format = "f"))
+  if(is.list(x) && any(idx <- vapply(x, inherits, NA, "difftime")))
+  {
+    x[idx] <- lapply(x[idx], function(xx) paste(trimws(formatC(unclass(xx), digits = digits, format = "f")), units(xx)))
+  }
   if(length(x) == 1) return(paste0(x))
 
   parens <- get_attr(x, "parens", c("", ""))